backend/requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
python-decouple
pandas
celery
redis
gunicorn
pandas
openpyxl